This make float frameskipping work actually!
Thanks cyclonmaster for the idea!

+++ Core/Config.h
@@ -72,7 +72,7 @@ public:
 #endif
 	bool bStretchToDisplay;
 	bool bVSync;
-	int iFrameSkip;
+	float iFrameSkip;
 
 	int iWindowX;
 	int iWindowY;
